Add start option to post build
Fixes #146
