+++ src/onPreBuild.test.js
@@ -1,0 +1,21 @@
+describe('onPreBuild', () => {
+  it.todo('installs cypress')
+
+  it.todo('skips when pre build tests are disabled')
+
+  it.todo('serves the publish folder')
+
+  it.todo('waits for the specified url to be ready before running tests')
+
+  it.todo('kills the process when tests are complete')
+
+  describe('recording', () => {
+    it.todo('records the test run when enabled and record key is specified')
+    it.todo('does not record the test run when record key is missing')
+    it.todo('does not record when recording is disabled')
+  })
+
+  it.todo('runs the cypress tests')
+
+  it.todo('processes the cypress test results')
+})
