+++ src/onSuccess.test.js
@@ -1,0 +1,15 @@
+describe('onSuccess', () => {
+  it.todo('skips when on success tests are disabled')
+
+  it.todo('errors when deploy url is not set')
+
+  describe('recording', () => {
+    it.todo('records the test run when enabled and record key is specified')
+    it.todo('does not record the test run when record key is missing')
+    it.todo('does not record when recording is disabled')
+  })
+
+  it.todo('runs the cypress tests')
+
+  it.todo('processes the cypress test results')
+})
